Add hidden tag to search extension
This was a suggestion from @mattwire that we hide the extension from the UI until the styling dependency is resovled.

It seems reasonable to me as this is primarily in the civicrm-core repo for code management purposes and
making it too easy to enable before the styling is resolved could lead to confusion.

I would quite like to enable on buildkit builds just for dev visibility - but that is something to think
about