civix#175 - Add support for shim scripts
Overview
--------

(NOTE: For this description, I reference the term "API" in the general sense of a programmatic interface -- such as
a hook or file-naming convention. It is not specifically about CRUD/DB APIs.)

The `civix` code-generator provides support for additional coding-conventions -- ones which are handier than core's and
more amenable to code-generation.  For example, it autoloads files from `xml/Menu/*.xml` and `**/*.mgd.php`.  The
technique for implementing this traditionally relies on generating a lot of boilerplate.

This patch introduces a new construct ("a shim file") which allows boilerplate to be maintained more easily.  A "shim"
is like an adapter or polyfill that translates a new API (e.g.  `*.mgd.php`) to an older API (e.g.
`hook_civicrm_managed`).  The "shim" may start out as a file in civix (or even as a bespoke file in some module) - and
then be migrated into `civicrm-core`.

See: https://github.com/totten/civix/issues/175

Before
------

The civix templates generate a few files, such as `mymod.php` and `mymod.civix.php`.
A typical example looks like this:

```php
// mymod.php - Implement hook_civicrm_xmlMenu
require_once 'mymod.civix.php';
function mymod_civicrm_xmlMenu(&$all, $the, $params) {
  _mymod_civix_civicrm_xmlMenu($all, $the, $params);
}
```

and

```php
// mymod.civix.php - Implement hook_civicrm_xmlMenu
function _mymod_civix_civicrm_xmlMenu(&$all, $the, $params) {
  foreach (_mosaico_civix_glob(__DIR__ . '/xml/Menu/*.xml') as $file) {
    $files[] = $file;
  }
}
```

These two files are managed differently: `mymod.php` is owned by the developer, and they may add/remove/manage the
hooks in this file.  `mymod.civix.php` is owned by `civix` and must be autogenerated.

This structure allows `civix` (and any civix-based extension) to take advantage of new coding-convention
immediately. However, it comes with a few pain-points:

* If you want to write a patch for `_mymod_civix_civicrm_xmlMenu`, the dev-test-loop requires several steps.
* If civix needs to add a new `hook_civicrm_foo`, then the author must manually create the stub
  function in `mymod.php`. `civix` has documentation (`UPGRADE.md`) which keeps a long list of stubs that must
  be manually added.
* If civix needs an update to its `hook_civicrm_foo`, then the author must regenerate `mymod.civix.php`.
* If civix's spin on `hook_civicrm_xmlMenu` becomes widespread, then the `xmlMenu` boilerplate is duplicated
  across many extensions.

After
-----

When loading an extension, Civi will look for files named `shims/*.shim.php`. Each file follows this pattern:

```php
return function(\CRM_Extension_ShimApi $shimApi) {
  // echo "This is " . $shimApi->longName . "!\n";
  \Civi::dispatcher()->addListener("hook_civicrm_xmlMenu", function($e) use ($shimApi) {
    ...
  });
}
```

The shim-file is a plain PHP file that can be debugged/copied/edited verbatim, and it can register for hooks on its
own.  The code is no longer a "template", and it doesn't need to be interwoven between `mymod.php` and
`mymod.civix.php`.

There is a presumption that all shims will be loaded.  However, if the equivalent functionality has been migrated into
`civicrm-core`, then `civicrm-core` may veto the shim-file.

Comments
--------

After this is merged, we can add some actual shims.

If it works out after the first RC, then I'd be fairly tempted to backport (eg to an ESR), but that's not a deal-breaker.