(REF) regen.sh - Remove unusual handling of `zipcodes.mysql`
Overview
--------

This is a small refactoring in service of the broader goal of simplifying
the sample-data generation and incorporating it into the installer (where it
can be localized/configured and where we don't have to commit any blobs).

The current process revolves around `regen.sh` and `civicrm_generated.mysql`.

Before
------

`regen.sh` builds an example DB and dumps the content into `civicrm_generated.mysql`.

As part of this, it loads `zipcodes.mysql` into the example DB,
drops/creates a table `zipcodes`, then runs some setup work,
and then re-drops `zipcodes`.

This is only safe on an isolated database. If you're trying to run it on
a pre-existing database created by `$cms`, then you don't know if there's going
to be a conflict on `zipcodes`.

After
-----

It doesn't drop/create/drop the `zipcodes` table. It just reads from the file.

Comment
-------

The main change is fairly small.  It looks big because `zipcodes.mysql` as
to be reformatted as `zipcodes.json`.