[REF] remove unnecessary variable variables
For a brief but terrifying spell this arvo I delved into this function to try to understand
the purpose of 'preserveDBName'. Obviously I ran screaming from the room but I thought I could disarm these
few lines to slighly file down the fangs of the beast.

Code is heavily covered by tests