Update flexmailer release information
I went to look at https://docs.civicrm.org/dev/en/latest/extensions/info-xml/ and found the example to
be out of date. I thought we could replace with flexmailer -but we are still calling it 'alpha' so
I updated the basic values for when it goes first ships with core (well I guess it will ship in 5.28 but
this will hit 5.29). Open question as to whether we keep incrementing the version
but I don't think it's urgent