[REF] Reduce interaction between dedupe code and createProfileContact
Ideally we want to simply call apiv4 from the deduper. However, it seems to take a bit of analysis to determine
if anything is done in there that might have some impact. This takes one step in that direction.  I copied the full function over
and then remove all the parts that are unreachable due to the passed in params or meaningless due to us not having Billing
fields. I think the Profile hooks are overkill - the merge hooks and contact hooks are being run and no profile is being submitted