[REF] Simplify location metadata handling in Export class
The metadata array that is passed into getTransformedFieldValue is just a subset of the class metadata. It is
- keyed slightly differently
- only location fields

Obviously passing this around when we can access it directly is silly and it was as a consquence as this code having
started off incredibly toxic and a slow incremental cleanup process that it was done this way. The cleanup had
not reached that array.

This has excellent test cover with multiple variations of phones & Ims tested along with all entities. These exposed
a tangental improvement (not replacing the word 'No' with '' ) Email On Hold.