[REF] Make explicit what we are doing with 'values' in this code
I've kept this really small so it can be verified by someone unfamiliar with this code.

Basically if you look you can see the only place 'values' is used is to determine if is_email_receipt is set.
That means everything else in  is unused & ergo we don't need to set it / can throw it away.

Obviously the next step is to extract all the code that determines  into a single place and say
goodbye to  altogether