Move function to delete merged contacts to the Merger class and include all instances of the pair
The PrevNextCache class is a bit of a mess. It has 2 roles
1) be a cache
2) provide functionality specific to the Merger class

I am working to move this second functionality to the merger class and also to remove the cacheKey
filter on it. The cacheKey pertains to a particular search set. However if 2 people are deduping different search
sets and one person merges a contact pair they should not remain available in the other person's result set.

(Likewise, and to follow, if 2 contacts have a conflict we should update all search sets with that information,
not just the one that was actively being deduped)