[REF] Move handling of form elements back to the Form
This toxic function getRowsElementsAndInfo does things for the form layer but it's also heavily used
in the BAO dedupeProcess. Ideally we want the things that are onnly being done to support (one specific) form
to sit on that form. This does that just for one of the 2 arrays it creates of form elements.

The rel_table_elements array is just an array that looks like

[
  ['checkbox',  'move_rel_contributions'],
  ['checkbox, 'move_rel_activities'],
]

The rel_tables array is an array of metadata keyed by eg. move_rel_contributions.

Ergo this can be done easily on the form layer with the data it already has