Reduce calls to CRM_Member_PseudoConstant::membershipType
This function defaults to not using the cached version so that's a pretty good argument for making a more
active effort to deprecate it. The createMembership function flushes it so we don't need to do that
so much in tests now either. I also included swapping status over in the test