[REF] Extract function to get locations to merge, rename 'operation' to is_replace'
This is a preliminary refactor. I have a bug to fix in this code but cannot yet make sense of it. The mergeHandler class
is simply an object to refactor the functions onto. Much of this code is hard to work with as the use of static functions
necessitates a lot of compilation and compiling so the process of cleaning it up involves extracting functionns to this
new class (which may one day replace the Merger class). As it is a refactoring process the functions reflect the old code
more than the ideal code. However, it provides the change to give some documentation regarding the locBlock array
and also to rename the confusing field 'operation' to 'is_replace'