Remove main PaymentExpress class
I recently identified that this processor had been broken for a long time & removed the broken IPN class

At that time I didn't remove the main class as I was unsure how to deal with any legacy transactions related to it.

However, @mattwire pointed out he had done work earlier to make CiviCRM cope with not-present processor classes and,
on testing, I was unable to trigger any errors after setting up a contribution page & an event page & making a payment
and then deleting the class. I couldn't edit the contribution, but with appropriate in app messaging and no fatal
error.