[REF] Stop passing transaction to completeTransaction
Transaction is an optional parameter for completeOrder and it would only make sense to pass it in
it the transaction were carrying some information.

However, in each of these cases no action has been taken since it was initiated to passing in NULL
is equally good.

(note this means a next step is to remove it from completeOrder signature too)