[NFC] [Test] Intial testing on event payment forms.
I'm trying to work on improving our testing of payment forms. I started with 3 event forms but there were
enough challenges that for this commit I'm just adding tests on one form. As noted in the code comments
it makes sense for this test to be in the extension but I want to work through the challenges on the
other forms before finalising any helper functions to the point where they are available to extensions.

This is actually a bit of a break through as it's the first time we have testing on a form flow - ie
submitting the first form and then the second. It was quite painful