Convert Manual processor to use the propertyBag
We encourage processors to cast to the propertyBag where it can meet their needs.

However the forms need to be altered to cope with the possibility this has happened. We can do this
by having them not attempt to re-use the params they pass to doPayment for other purposes. As
https://github.com/civicrm/civicrm-core/pull/17507 highlights this function continues to
use $membershipParams in the outer functionn - unaware that within it's depths a processor has
converted it to a propetyBag - in that PR it is then cast back but one functions
standardisation is another functions 'tampering' - any code calling this function should
not expect $params['is_recur'] to be replaced with $params['isRecur'].

In fact the only reason everything is being passed from pillar to post is that is how the code worked going back to
2.x & we are too scared to undo all those pass-by-references.