dev/financial#135 Fix Authorize.net to implement doPayment not doDirectPayment
We have recommended processors implement doPayment and NOT doDirectPayment or doTransferPayment for a long
time but have not modelled this in core, leading to the wrong code being copied & propagaged into
extensions. As Karing outlined here https://lab.civicrm.org/dev/financial/-/issues/135
we need to ensure exceptions are thrown - that is already done. We also need to

1) ensure this->component is set for the return urls
2) handlle zero amounts
3) ensure payment_status_id is set
4) rename the function