[REF] Clean up return variables on updateContributionStatus, updatePendingOnlineContribution
The return variable from updatePendingOnlineContribution is never used - ergo don't return it. This means
the return value from updateContributionStatus is also never used - ergo - you get it. In addition don't call the function
with empty variables - I've moved the check as to whether they are empty into the calling function - although likely
in a second pass the checks will be removed or change again