Move tax handling from line item api to BAO to make it available from apiv4
This implementation has some limitations. I only address one in this PR - removing the rounding - as
the focus of the PR is the move.

The rounding from all save layers was previously removed but it was reverted when it was eroneously believed
to have caused a bug. The bug turned out to be https://github.com/civicrm/civicrm-core/pull/18297