[REF] Simplify array construction
In trying to interpret https://github.com/civicrm/civicrm-core/pull/17992

I realised that the code copies values from the dao to an array, copies the array
to another array, unsets most of the variables and then uses that second array.

This simplifies.

The easiest way to work through this code is in a debugger running

testProcessMembershipUpdateStatus