dev/financial#149 Add permission = false to apiv4 call to address regression
This retrieves payment processors using FALSE for get permissions. I was able to replicate the bug in the test.

I also updated one line that I passed through when stepping through the test that related to the addActivity.
It just passes in activityType rather than leaving addActivity to figure that out