[REF] Start the process of separating the search action from the participant form.
A lot code nastiness stems from the overloading of the participant registration form with the search task to register
multiples. A more logical structure would be some shared functionality but separate classes. This starts that
process, but with only the one property now separated to be a class property rather than a derived one. Later efforts
can disentangle which functionality belongs to which class