[REF] Minor code extraction
This extracts code to the MergeHandler class for deduping locations. We have a LOT of
test cover here with all the scenarios in jobTest::testBatchMergesAddresses & other functions
using that same data provider.

The goal is to refactor the code onto the MergeHandler class, with a view to
doing less variable passing & more property-using to help clean it up -
in this case it's already visible that 2 variables can be derived in the mergeHandler: