CRM_Core_Resources::addBundle() - Fix handling of array inputs
(This fixes a small bug in a new function added during this release -- part of #18247.)

The signature of `addBundle()` optionally accepts an array/iterable -- if
given, then it should add all the items from the array. For example:

```php
Civi::resources()->addBundle(['foo', 'bar']);
```

Before
------

It adds `foo` but then bails out on `bar`.

After
-----

It adds both `foo` and `bar`.