Duplicate preProcessCommmon
The preProcessCommon is a gnarly bit of the task subsystem. If it were not
a static function it would be much easier to identify & cleanup the common elements.

It is called from 2 places outside of the inherited class structure - here & in export.

I propose to
1) take a copy of it here & then clean it up on this class in a series of follow ups
2) create subclasses per component of the export class rather than the weird query
parsing.

We have had mixed results with the 'duplicate & clean' approach - notably because
sometimes it has been 'duplicate & dump' - ie copy code to other places in order to
clean up the 'main place' & leave the other place in a mess. But I'm going to proceded
with further cleanup of this duplicated code on this function - I suspect only a line or
2 are really relevant here