[Ref] Extract getFormValues
Digging into this is seems the  is always an object that inherits from this class

It is called statically primarily because the export task does not have
subclasses per form & does not know what variant it's passing in. However, we can
move functions onto the CRM_Contact_Form_Task class as non-static functions
and call them form form-> however unintuitive that seems.

I am trying to clean up the structure such that we can move on from the
static method but decided I should tackle this function a little first.

To test try doing an export from any search & this is called in static mode
(same goes for any task but the path is less ambiguous)