dev/core#2057 Remove select query that never or almost never finds something.
When it comes to quey efficiency it's good to do an extra select if it saves an update/insert
but not if it never does. To test the theory that this 'never does' I put
removing this select through the entire test suite. No tests failed, meaning
it was unnecessary 100% of the time. Adding the catch ensures
it will still succeed if it the row exists but we have moved our query conservation
from doing an extraneous select in 99.9% of cases to an extraneous update in .1% of cases

https://lab.civicrm.org/dev/core/-/issues/2057