dev/core#2046 Ensure location BAOs have create actions, deprecate add, simplify block code
Per https://lab.civicrm.org/dev/core/-/issues/2046 we have a goal (long term) to consolidate create & add functions into one function - with the functionality moved from add into create & add becoming a deprecated wrapper around create.

This makes that change on the IM & OpenID in line with the previous change on Phone & Email, leaving just Address & Website which are
both a bit trickier. The code in Block runs on the 4 that are updated so this simplifies & confirms all 4 are calling
functions that handle primary. We could switch to an apiv4 call - I just want to be a bit more sure we've handlled primary there
& also that we look into Address & website