dev/core#1790 - Contact Card - Email Links
Alternate to https://github.com/civicrm/civicrm-core/pull/17517

The current PR stalled on the issue that passing 'email' to the email task could load up
the wrong contact. This gets around that by doing a separate (but hopefully quick) lookup
to grab the email id. I also made it dependent on the url having is_show_email_task=1 in
it. I went back & forwards on that vs is_backoffice