dev/core#1393 - distmaker - Skip new file "vendor/bin/cssmin"
Overview
----------------------------------------

This fixes the broken distmaker builds in `master`.

The file `vendor/bin/cssmin` was recently added by way of `composer-compile-lib => cssmin`.
We don't need it in the tarball, and it's a symlink, so trying to include breaks some
consumers (like in https://lab.civicrm.org/dev/core/-/issues/1393).

Technical Details
----------------------------------------

* Before (last release): `vendor/bin/cssmin` does not exist
* Before (current master): `vendor/bin/cssmin` is created as symlink, copied to tarball
* After (this commit): `vendor/bin/cssmin` is created on dev-builds but omitted from tarball