dev/core#2046 Fix blockDelete to delete while ensuring is_primary is valid
The del function on each of the location DAO ensures that after a delete at least one phone (etc) is
marked is_primary. This blockDelete does not do that & per the test it's possible to call this
in such a way that no addresses are primary. Whether this actually happens is a bit unknown.

However, this cleans up block delete in a way where we can demonstrate the tests work