dev/core#2079 Eliminate unused query on CRM_Core_BAO_CustomQuery::_construct
This cleans up unnecessary queries in Contact.get api when custom data is involved. As mentioned _options is deprecated
but still used - more accurately the work wasn't done at the time to determine whether it was.

This time I've gone through and clarified all the places it potentially might have been used

https://github.com/civicrm/civicrm-core/pulls?q=+is%3Apr+dev%2Fcore%232079+