dev/financial#152 Remove unreachable code
This code cann only be reached if txnType is recurring_payment. Given that we can look
at the earlier switch and see that in that case first must be TRUE or it would have
already returned. Ergo this code is unreachable - and would have been
since the earlier code switched to using the api to handle non-firsts