Fix campaign_id handling for batch entry
Fixes a bug whereby campaign_id is not updated on batch entry if it has been added to the profile.

The underlying issue appears to be an  attack of copy and paste. I tidied up the variable
in both the places that call the processMembership function such that it is passed in
as an array of pass-through params. The function can be simplified & eventually removed.